[改善]新たにsharesコントローラーを作成し、question_answersコントローラーからshareアクションを移動
