+++ app/assets/stylesheets/question_answer/index.css
@@ -125,7 +125,6 @@
   position: relative;
   width: 580px;
   margin-top: 5px;
-  outline: 1px solid #000;
 }
 
 .qa-index-item__record {
