+++ app/helpers/shares_helper.rb
@@ -1,0 +1,2 @@
+module SharesHelper
+end
