+++ config/routes.rb
@@ -1,19 +1,21 @@
 Rails.application.routes.draw do
+  root 'groups#index'
   devise_for :users, controllers: {
     registrations: 'users/registrations'
   }
-  root 'groups#index'
   resources :groups, only: [:index, :create, :update, :destroy] do
-    resources :question_answers do
+    resources :question_answers, except: [:show] do
       collection do
         get 'review', 'change_date' # change_dateは挙動確認用。アプリリリース時には削除
       end
       member do
         get 'search_category'
-        post 'share'
         patch 'reset', 'remove'
       end
     end
+  end
+  resources :question_answers, only: [:show] do
+    resources :shares, only: [:create]
   end
   resources :options, only: [:edit, :update]
   patch '/repetition_algorithms', to: 'repetition_algorithms#update'
