+++ spec/requests/shares_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "Shares", type: :request do
+
+end
